LayoutTests/inspector/console-dir.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="console-tests.js"></script>
<script>

function doit()
{
    console.dir(document);
    console.dir(["test1", "test2"]);
    console.dir(document.childNodes);

    dumpConsoleMessages();
}

</script>
</head>

<body onload="onload()">
<p>
Tests that console logging dumps proper messages.
</p>

</body>
</html>
