LayoutTests/inspector/console-dir-expected.txt

CONSOLE MESSAGE: line 9: [object HTMLDocument]
CONSOLE MESSAGE: line 10: test1,test2
CONSOLE MESSAGE: line 11: [object NodeList]
Tests that console logging dumps proper messages.

console-dir.html:9HTMLDocument
console-dir.html:10Array
console-dir.html:11NodeList
